+++ src/main/java/me/wonk2/utilities/enums/DFType.java
@@ -1,0 +1,13 @@
+package me.wonk2.utilities.enums;
+
+public enum DFType {
+	TXT,
+	NUM,
+	SND,
+	LOC,
+	ITEM,
+	POT,
+	VAR,
+	LIST,
+	ANY // This is only used internally for parameter types.
+}
