+++ src/main/java/me/wonk2/utilities/enums/Scope.java
@@ -1,0 +1,7 @@
+package me.wonk2.utilities.enums;
+
+public enum Scope {
+	GLOBAL,
+	LOCAL,
+	SAVE
+}
