--- utilities/enums/DFType.java
@@ -1,13 +1,0 @@
-package me.wonk2.utilities.enums;
-
-public enum DFType {
-	TXT,
-	NUM,
-	SND,
-	LOC,
-	ITEM,
-	POT,
-	VAR,
-	LIST,
-	ANY // This is only used internally for parameter types.
-}
