--- utilities/enums/Scope.java
@@ -1,7 +1,0 @@
-package me.wonk2.utilities.enums;
-
-public enum Scope {
-	GLOBAL,
-	LOCAL,
-	SAVE
-}
